backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.46
pydantic==1.10.7
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.28.2
PyJWT==2.7.0
